+++ presto-orc/src/main/java/com/facebook/presto/orc/stream/LongStreamV2.java
@@ -390,7 +390,7 @@ public class LongStreamV2
     }
 
     @Override
-    public void skip(int items)
+    public void skip(long items)
             throws IOException
     {
         while (items > 0) {
